claudeflow/requirements.txt
# ClaudeFlow依赖包

# YAML配置文件支持
pyyaml>=6.0

# 彩色终端输出
colorama>=0.4.6

# MCP服务器支持
mcp[cli]>=1.4.0

# 代码分析核心(如果需要的话)
# code2flow>=2.5.0